Include build-script to make use of rollup config
As there already is a rollup.config.js included in this project I expected that there would be a build-script too which makes use of this file. So far rollup is only used for the docs:build and checksize task. Including this new script would help beginners to make it more easier to actually use the component later on. Throughout the docs roolup is mentioned first when talking about building your component so I think it's fair to include a step to build the component. An alternative would be to also include a webpack config and also step for that but I'd argument for rollup to keep the starter project simple